+++ src/filters/vibrance.shaders.ts
@@ -1,0 +1,16 @@
+export const fragmentSource = `
+  precision highp float;
+  uniform sampler2D uTexture;
+  uniform float uVibrance;
+  varying vec2 vTexCoord;
+  void main() {
+    vec4 color = texture2D(uTexture, vTexCoord);
+    float max = max(color.r, max(color.g, color.b));
+    float avg = (color.r + color.g + color.b) / 3.0;
+    float amt = (abs(max - avg) * 2.0) * uVibrance;
+    color.r += max != color.r ? (max - color.r) * amt : 0.00;
+    color.g += max != color.g ? (max - color.g) * amt : 0.00;
+    color.b += max != color.b ? (max - color.b) * amt : 0.00;
+    gl_FragColor = color;
+  }
+`;
